# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Codeplane API
  version: 1.0.0
  description: API for Codeplane - a jj-native development platform
servers:
  - url: https://api.codeplane.app
paths:
...
  /repos/{owner}/{repo}/workspaces:
    get:
      summary: GET /repos/{owner}/{repo}/workspaces
      description: List first-class workspaces for the repository.
      operationId: get_repos__owner___repo__workspaces
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          description: Opaque cursor from a previous response's next_cursor field
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
            maximum: 100
      responses:
        '200':
          description: List of workspaces
          headers:
            Link:
              description: RFC 8288 pagination links with cursor-based URLs.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
                  next_cursor:
                    type: string
                    nullable: true
                    description: Opaque cursor for fetching the next page. Null when there are no more results.
    post:
      summary: POST /repos/{owner}/{repo}/workspaces
      description: Create or resume a cloud workspace.
      operationId: post_repos__owner___repo__workspaces
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Optional workspace name.
                snapshot_id:
                  type: string
                  description: Restore the workspace from a saved snapshot.
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
  /repos/{owner}/{repo}/workspaces/{id}:
    get:
      summary: GET /repos/{owner}/{repo}/workspaces/{id}
      description: Get a workspace.
      operationId: get_repos__owner___repo__workspaces__id_
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
    delete:
      summary: DELETE /repos/{owner}/{repo}/workspaces/{id}
      description: Delete a workspace and its backing cloud VM.
      operationId: delete_repos__owner___repo__workspaces__id_
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Workspace deleted
  /repos/{owner}/{repo}/workspaces/{id}/ssh:
    get:
      summary: GET /repos/{owner}/{repo}/workspaces/{id}/ssh
      description: Get SSH connection info for a workspace.
      operationId: get_repos__owner___repo__workspaces__id__ssh
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSH connection info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceSSHConnectionInfo'
  /repos/{owner}/{repo}/workspaces/{id}/suspend:
    post:
      summary: POST /repos/{owner}/{repo}/workspaces/{id}/suspend
      description: Suspend a running workspace.
      operationId: post_repos__owner___repo__workspaces__id__suspend
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Suspended workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
  /repos/{owner}/{repo}/workspaces/{id}/resume:
    post:
      summary: POST /repos/{owner}/{repo}/workspaces/{id}/resume
      description: Resume a suspended workspace.
      operationId: post_repos__owner___repo__workspaces__id__resume
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resumed workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
  /repos/{owner}/{repo}/workspaces/{id}/fork:
    post:
      summary: POST /repos/{owner}/{repo}/workspaces/{id}/fork
      description: Fork a running workspace into a new derived workspace.
      operationId: post_repos__owner___repo__workspaces__id__fork
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Optional name for the forked workspace.
      responses:
        '201':
          description: Forked workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
  /repos/{owner}/{repo}/workspaces/{id}/snapshot:
    post:
      summary: POST /repos/{owner}/{repo}/workspaces/{id}/snapshot
      description: Create a reusable snapshot from a workspace.
      operationId: post_repos__owner___repo__workspaces__id__snapshot
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Snapshot name.
      responses:
        '201':
          description: Snapshot created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceSnapshot'
  /repos/{owner}/{repo}/workspaces/{id}/stream:
    get:
      summary: GET /repos/{owner}/{repo}/workspaces/{id}/stream
      description: SSE endpoint for workspace lifecycle events.
      operationId: get_repos__owner___repo__workspaces__id__stream
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
  /repos/{owner}/{repo}/workspace-snapshots:
    get:
      summary: GET /repos/{owner}/{repo}/workspace-snapshots
      description: List reusable workspace snapshots for the repository.
      operationId: get_repos__owner___repo__workspace_snapshots
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          description: Opaque cursor from a previous response's next_cursor field
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
            maximum: 100
      responses:
        '200':
          description: List of workspace snapshots
          headers:
            Link:
              description: RFC 8288 pagination links with cursor-based URLs.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkspaceSnapshot'
                  next_cursor:
                    type: string
                    nullable: true
                    description: Opaque cursor for fetching the next page. Null when there are no more results.
    post:
      summary: POST /repos/{owner}/{repo}/workspace-snapshots
      description: Create a snapshot template from an existing workspace.
      operationId: post_repos__owner___repo__workspace_snapshots
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - workspace_id
              properties:
                workspace_id:
                  type: string
                  description: Source workspace for the snapshot.
                name:
                  type: string
                  description: Snapshot name.
      responses:
        '201':
          description: Snapshot created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceSnapshot'
  /repos/{owner}/{repo}/workspace-snapshots/{id}:
    get:
      summary: GET /repos/{owner}/{repo}/workspace-snapshots/{id}
      description: Get a workspace snapshot.
      operationId: get_repos__owner___repo__workspace_snapshots__id_
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace snapshot details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceSnapshot'
    delete:
      summary: DELETE /repos/{owner}/{repo}/workspace-snapshots/{id}
      description: Delete a workspace snapshot.
      operationId: delete_repos__owner___repo__workspace_snapshots__id_
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Workspace snapshot deleted
...
  /repos/{owner}/{repo}/workspace/sessions/{id}:
    get:
      summary: GET /repos/{owner}/{repo}/workspace/sessions/{id}
      description: Get a workspace session.
      operationId: get_repos__owner___repo__workspace_sessions__id_
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceSession'
  /repos/{owner}/{repo}/workspace/sessions/{id}/ssh:
    get:
      summary: GET /repos/{owner}/{repo}/workspace/sessions/{id}/ssh
      description: Get SSH connection info for a workspace session.
      operationId: get_repos__owner___repo__workspace_sessions__id__ssh
      tags:
        - Workspaces
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSH connection info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceSSHConnectionInfo'
...
components:
  schemas:
...
    Workspace:
      type: object
      properties:
        id:
          type: string
          description: Unique workspace identifier
          example: ws_abc123
        repository_id:
          type: integer
          format: int64
          description: Repository this workspace belongs to
        user_id:
          type: integer
          format: int64
          description: User who owns the workspace
        name:
          type: string
          description: Human-readable workspace name
          example: primary
        status:
          type: string
          enum: [pending, starting, running, suspended, stopped, failed]
          description: Current workspace status
          example: running
        is_fork:
          type: boolean
          description: Whether this workspace was forked from another workspace
        parent_workspace_id:
          type: string
          nullable: true
          description: Source workspace ID for forked workspaces
        freestyle_vm_id:
          type: string
          description: VM identifier
          example: vm-xyz789
        persistence:
          type: string
          enum: [sticky, ephemeral, persistent]
          description: Workspace persistence level
          example: sticky
        ssh_host:
          type: string
          nullable: true
          description: SSH connection host
          example: vm-xyz789@vm-ssh.codeplane.app
        snapshot_id:
          type: string
          nullable: true
          description: Snapshot template used to create this workspace
          example: snap_def456
        idle_timeout_seconds:
          type: integer
          description: Seconds of inactivity before auto-suspend
          example: 1800
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        suspended_at:
          type: string
          format: date-time
          nullable: true
          description: When the workspace was suspended (null if running)
...